--- src/Traits/Styling/Configuration/ColumnSelectStylingConfiguration.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Rappasoft\LaravelLivewireTables\Traits\Styling\Configuration;
-
-trait ColumnSelectStylingConfiguration
-{
-    public function setColumnSelectButtonAttributes(array $attributes = []): self
-    {
-        $this->columnSelectButtonAttributes = [...$this->columnSelectButtonAttributes, ...$attributes];
-
-        return $this;
-    }
-
-    public function setColumnSelectMenuOptionCheckboxAttributes(array $attributes = []): self
-    {
-        $this->columnSelectMenuOptionCheckboxAttributes = [...$this->columnSelectMenuOptionCheckboxAttributes, ...$attributes];
-
-        return $this;
-    }
-}
