--- src/Traits/Styling/Configuration/ToolsStylingConfiguration.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Rappasoft\LaravelLivewireTables\Traits\Styling\Configuration;
-
-trait ToolsStylingConfiguration
-{
-    public function setToolsAttributes(array $toolsAttributes = []): self
-    {
-        $this->setCustomAttributes(propertyName: 'toolsAttributes', customAttributes: $toolsAttributes);
-
-        return $this;
-    }
-
-    public function setToolBarAttributes(array $toolBarAttributes = []): self
-    {
-        $this->setCustomAttributes(propertyName: 'toolBarAttributes', customAttributes: $toolBarAttributes);
-
-        return $this;
-    }
-}
