+++ src/Traits/Styling/HasLoadingPlaceholderStyling.php
@@ -2,14 +2,8 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits\Styling;
 
-use Rappasoft\LaravelLivewireTables\Traits\Styling\Configuration\LoadingPlaceholderStylingConfiguration;
-use Rappasoft\LaravelLivewireTables\Traits\Styling\Helpers\LoadingPlaceholderStylingHelpers;
-
 trait HasLoadingPlaceholderStyling
 {
-    use LoadingPlaceholderStylingConfiguration,
-        LoadingPlaceholderStylingHelpers;
-
     protected array $loadingPlaceHolderAttributes = [];
 
     protected array $loadingPlaceHolderIconAttributes = [];
@@ -19,4 +13,61 @@ trait HasLoadingPlaceholderStyling
     protected array $loadingPlaceHolderRowAttributes = [];
 
     protected array $loadingPlaceHolderCellAttributes = ['class' => '', 'default' => true];
+
+    public function getLoadingPlaceholderAttributes(): array
+    {
+        return $this->getCustomAttributes(propertyName: 'loadingPlaceHolderAttributes', default: true, classicMode: true);
+
+    }
+
+    public function getLoadingPlaceHolderIconAttributes(): array
+    {
+        return $this->getCustomAttributes(propertyName: 'loadingPlaceHolderIconAttributes', default: true, classicMode: true);
+
+    }
+
+    public function getLoadingPlaceHolderWrapperAttributes(): array
+    {
+        return $this->getCustomAttributes(propertyName: 'loadingPlaceHolderRowAttributes', default: true, classicMode: true);
+    }
+
+    public function getLoadingPlaceHolderRowAttributes(): array
+    {
+        return $this->getCustomAttributes(propertyName: 'loadingPlaceHolderRowAttributes', default: true, classicMode: true);
+    }
+
+    public function getLoadingPlaceHolderCellAttributes(): array
+    {
+        return $this->getCustomAttributes(propertyName: 'loadingPlaceHolderCellAttributes', default: true, classicMode: true);
+
+    }
+
+    public function setLoadingPlaceHolderAttributes(array $attributes): self
+    {
+        $this->setCustomAttributes('loadingPlaceHolderAttributes', [...$this->getCustomAttributes(propertyName: 'loadingPlaceHolderAttributes', default: false, classicMode: true), ...$attributes]);
+
+        return $this;
+    }
+
+    public function setLoadingPlaceHolderIconAttributes(array $attributes): self
+    {
+        $this->setCustomAttributes('loadingPlaceHolderIconAttributes', [...$this->getCustomAttributes(propertyName: 'loadingPlaceHolderIconAttributes', default: false, classicMode: true), ...$attributes]);
+
+        return $this;
+    }
+
+    public function setLoadingPlaceHolderRowAttributes(array $attributes): self
+    {
+        $this->setCustomAttributes('loadingPlaceHolderRowAttributes', [...$this->getCustomAttributes(propertyName: 'loadingPlaceHolderRowAttributes', default: false, classicMode: true), ...$attributes]);
+
+        return $this;
+    }
+
+    public function setLoadingPlaceHolderWrapperAttributes(array $attributes): self
+    {
+        $this->setCustomAttributes('loadingPlaceHolderRowAttributes', [...$this->getCustomAttributes(propertyName: 'loadingPlaceHolderRowAttributes', default: false, classicMode: true), ...$attributes]);
+
+        return $this;
+    }
+
 }
