+++ src/Traits/Styling/HasReorderStyling.php
@@ -2,13 +2,38 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits\Styling;
 
-use Rappasoft\LaravelLivewireTables\Traits\Styling\Configuration\ReorderStylingConfiguration;
-use Rappasoft\LaravelLivewireTables\Traits\Styling\Helpers\ReorderStylingHelpers;
+use Livewire\Attributes\Computed;
 
 trait HasReorderStyling
 {
-    use ReorderStylingConfiguration,
-        ReorderStylingHelpers;
 
     protected array $reorderThAttributes = ['default' => true];
+
+    /**
+     * Used to get attributes for the <th> for Bulk Actions
+     *
+     * @return array<mixed>
+     */
+    #[Computed]
+    public function getReorderThAttributes(): array
+    {
+        return $this->reorderThAttributes ?? ['default' => true];
+    }
+
+    #[Computed]
+    public function hasReorderThAttributes(): bool
+    {
+        return $this->getReorderThAttributes() != ['default' => true];
+    }
+
+    /**
+     * Used to set attributes for the <th> for Reorder Column
+     */
+    public function setReorderThAttributes(array $reorderThAttributes): self
+    {
+        $this->reorderThAttributes = [...$this->reorderThAttributes, ...$reorderThAttributes];
+
+        return $this;
+    }
+
 }
