--- src/Traits/Styling/Helpers/ToolsStylingHelpers.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace Rappasoft\LaravelLivewireTables\Traits\Styling\Helpers;
-
-use Illuminate\View\ComponentAttributeBag;
-use Livewire\Attributes\Computed;
-
-trait ToolsStylingHelpers
-{
-    #[Computed]
-    public function getToolsAttributes(): array
-    {
-        return $this->getCustomAttributes(propertyName: 'toolsAttributes', default: false, classicMode: false);
-    }
-
-    #[Computed]
-    public function getToolsAttributesBag(): ComponentAttributeBag
-    {
-        return $this->getCustomAttributesBagFromArray($this->getToolsAttributes());
-    }
-
-    protected function getToolBarAttributes(): array
-    {
-        return $this->getCustomAttributes(propertyName: 'toolBarAttributes', default: false, classicMode: false);
-    }
-
-    #[Computed]
-    public function getToolBarAttributesBag(): ComponentAttributeBag
-    {
-        return $this->getCustomAttributesBagFromArray($this->getToolBarAttributes());
-
-    }
-}
